moves issueticketform files to one subdir of tools
